requirements_backup.txt
absl-py==2.0.0
actionlib==1.14.0
angles==1.9.13
apturl==0.5.2
astroid==3.0.3
asttokens==2.4.1
astunparse==1.6.3
attrs==19.3.0
autobahn==17.10.1
Automat==0.8.0
autopep8==2.0.4
backcall==0.2.0
bcrypt==3.1.7
blinker==1.4
bondpy==1.8.6
breezy==3.0.2
Brlapi==0.7.0
cachetools==5.3.2
camera-calibration==1.17.0
camera-calibration-parsers==1.12.0
catkin==0.8.10
catkin-pkg==1.0.0
catkin-pkg-modules==1.0.0
catkin-tools==0.9.2
cbor==1.0.0
certifi==2019.11.28
chardet==3.0.4
charset-normalizer==3.3.2
clang==5.0
Click==7.0
colorama==0.4.3
colour==0.1.5
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
contourpy==1.1.1
controller-manager==0.20.0
controller-manager-msgs==0.20.0
cryptography==2.8
cupshelpers==1.0
cv-bridge==1.16.2
cycler==0.10.0
Cython==0.29.14
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
defusedxml==0.6.0
Deprecated==1.2.7
diagnostic-analysis==1.11.0
diagnostic-common-diagnostics==1.11.0
diagnostic-updater==1.11.0
dill==0.3.8
distro==1.4.0
distro-info==0.23+ubuntu1.1
docstring-to-markdown==0.15
docutils==0.16
dulwich==0.19.15
dynamic-reconfigure==1.7.3
empy==3.3.2
entrypoints==0.3
executing==2.0.1
ez-setup==0.9
fastimport==0.9.8
flake8==7.0.0
flatbuffers==23.5.26
fonttools==4.47.0
gast==0.4.0
gazebo-plugins==2.9.2
gazebo-ros==2.9.2
gencpp==0.7.0
geneus==3.0.0
genlisp==0.4.18
genmsg==0.6.0
gennodejs==2.0.2
genpy==0.6.15
google-auth==2.26.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
gpg==1.13.1
grpcio==1.60.0
h5py==3.1.0
httplib2==0.14.0
hyperlink==19.0.0
i3pystatus==3.35
idna==2.8
image-geometry==1.16.2
importlib-metadata==1.5.0
importlib-resources==6.1.1
incremental==16.10.1
interactive-markers==1.12.0
ipython==8.12.3
isort==5.13.2
jedi==0.19.1
joblib==1.3.2
joint-state-publisher==1.15.1
joint-state-publisher-gui==1.15.1
keras==2.10.0
Keras-Preprocessing==1.1.2
keyring==18.0.1
kiwisolver==1.0.1
language-selector==0.1
laser-geometry==1.6.7
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
libclang==16.0.6
louis==3.12.0
lz4==3.0.2+dfsg
macaroonbakery==1.3.1
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.1.2
matplotlib-inline==0.1.6
mccabe==0.7.0
meson==0.53.2
message-filters==1.16.0
more-itertools==4.2.0
mpi4py==3.0.3
netifaces==0.10.4
nose==1.3.7
numpy==1.17.4
oauthlib==3.1.0
olefile==0.46
opt-einsum==3.3.0
osrf-pycommon==2.0.2
packaging==23.2
pandas==2.0.3
paramiko==2.6.0
parso==0.8.3
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
platformdirs==4.2.0
pluggy==1.4.0
prompt-toolkit==3.0.43
protobuf==3.6.1
psutil==5.5.1
ptyprocess==0.7.0
pure-eval==0.2.2
py-ubjson==0.14.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycodestyle==2.11.1
pycryptodomex==3.6.1
pycups==1.9.73
pydocstyle==6.3.0
pydot==1.4.1
pyflakes==3.2.0
PyGithub==1.43.7
Pygments==2.3.1
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pylint==3.0.4
pymacaroons==0.13.0
PyNaCl==1.3.0
PyOpenGL==3.1.0
pyOpenSSL==19.0.0
pyparsing==2.4.6
pypng==0.0.20
PyQRCode==1.2.1
PyQt5==5.14.1
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pyqtgraph==0.13.3
pyRFC3339==1.1
pyserial==3.4
PySide2==5.15.2.1
PySide6==6.6.2
PySide6-Addons==6.6.2
PySide6-Essentials==6.6.2
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.7.3
python-debian==0.1.36+ubuntu1.1
python-gitlab==2.0.1
python-gnupg==0.4.5
python-lsp-jsonrpc==1.1.2
python-lsp-server==1.10.1
python-qt-binding==0.4.4
python-snappy==0.5.3
pytoolconfig==1.3.1
PyTrie==0.2
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
resource-retriever==1.12.7
roman==2.0.0
rope==1.12.0
rosbag==1.16.0
rosboost-cfg==1.15.8
rosclean==1.15.8
roscreate==1.15.8
rosdep==0.22.2
rosdep-modules==0.22.2
rosdistro==0.9.0
rosdistro-modules==0.9.0
rosgraph==1.16.0
rosinstall==0.7.8
rosinstall-generator==0.1.23
roslaunch==1.16.0
roslib==1.15.8
roslint==0.12.0
roslz4==1.16.0
rosmake==1.15.8
rosmaster==1.16.0
rosmsg==1.16.0
rosnode==1.16.0
rosparam==1.16.0
rospkg==1.5.0
rospkg-modules==1.5.0
rospy==1.16.0
rosserial-arduino==0.9.2
rosserial-client==0.9.2
rosserial-python==0.9.2
rosservice==1.16.0
rostest==1.16.0
rostopic==1.16.0
rosunit==1.15.8
roswtf==1.16.0
rqt-action==0.4.9
rqt-bag==0.5.1
rqt-bag-plugins==0.5.1
rqt-console==0.4.12
rqt-dep==0.4.12
rqt-graph==0.4.14
rqt-gui==0.5.3
rqt-gui-py==0.5.3
rqt-image-view==0.4.17
rqt-launch==0.4.9
rqt-logger-level==0.4.12
rqt-moveit==0.5.11
rqt-msg==0.4.10
rqt-nav-view==0.5.7
rqt-plot==0.4.13
rqt-pose-view==0.5.11
rqt-publisher==0.4.10
rqt-py-common==0.5.3
rqt-py-console==0.4.10
rqt-reconfigure==0.5.5
rqt-robot-dashboard==0.5.8
rqt-robot-monitor==0.5.15
rqt-robot-steering==0.5.12
rqt-runtime-monitor==0.5.10
rqt-rviz==0.7.0
rqt-service-caller==0.4.10
rqt-shell==0.4.11
rqt-srv==0.4.9
rqt-tf-tree==0.6.4
rqt-top==0.4.10
rqt-topic==0.4.13
rqt-web==0.4.10
rsa==4.9
rviz==1.14.20
scikit-learn==1.3.2
scipy==1.3.3
screen-resolution-extra==0.0.0
seaborn==0.13.1
SecretStorage==2.3.1
sensor-msgs==1.13.1
service-identity==18.1.0
shiboken2==5.15.2.1
shiboken6==6.6.2
simplejson==3.16.0
sip==4.19.21
six==1.14.0
smach==2.5.2
smach-ros==2.5.2
smclib==1.8.6
snowballstemmer==2.2.0
stack-data==0.6.3
systemd-python==234
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.10.0
tensorflow-gpu==2.10.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==1.1.0
tf==1.13.2
tf-conversions==1.13.2
tf2-geometry-msgs==0.7.7
tf2-kdl==0.7.7
tf2-py==0.7.7
tf2-ros==0.7.7
threadpoolctl==3.2.0
tomli==2.0.1
tomlkit==0.12.4
topic-tools==1.16.0
traitlets==5.14.1
Twisted==18.9.0
txaio==2.10.0
typing-extensions==4.10.0
tzdata==2023.4
u-msgpack-python==2.1
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36
ujson==5.9.0
unattended-upgrades==0.1
urllib3==1.25.8
urwid==2.0.1
vcstools==0.1.42
wadllib==1.3.3
wcwidth==0.2.13
werkzeug==3.0.1
whatthepatch==1.0.5
wrapt==1.11.2
wsaccel==0.6.2
wstool==0.1.18
xacro==1.14.17
xkit==0.0.0
yapf==0.40.2
zipp==1.0.0
zope.interface==4.7.1
